Fix the error with the model names
